Millores estructuració codi, tests
