+++ backend/app/routes/task.py
@@ -1,9 +1,6 @@
-from fastapi import APIRouter, Depends
-from sqlalchemy.orm import Session
+from fastapi import APIRouter
 
-from app.db import get_db
-from app.deps import get_current_verified_user
-from app.models import User
+from app.deps import CurrentVerifiedUser, DbSession
 from app.schemas import TaskResponse
 from app.services import task_service
 
@@ -13,15 +10,15 @@ router = APIRouter()
 @router.get("/task")
 def get_task(
     category_code: str,
-    db: Session = Depends(get_db),
-    current_user: User = Depends(get_current_verified_user),
+    current_user: CurrentVerifiedUser,
+    db: DbSession,
 ) -> TaskResponse:
     """
     Retorna la propera tasca per a un usuari utilitzant el servei task_service
     Args:
-        db: sessió SQLAlchemy
         category_code: codi de la categoria (e.g. "correccio", "reformulacio")
-        current_user: usuari autenticat i verificat
+        current_user: usuari autenticat i verificat (injectat via dependència)
+        db: sessió SQLAlchemy
     Returns:
         TaskResponse: objecte amb el prompt, les dues respostes i el token
     """
