+++ backend/pyproject.toml
@@ -9,6 +9,7 @@ dependencies = [
     "alembic>=1.13",
     "psycopg[binary]>=3.1",
     "pydantic-settings>=2.0",
+    "email-validator>=2.0",
     "argon2-cffi>=23.1",
     "pyyaml>=6.0",
     "numpy==2.5.0",
